elsmd conversion: first finished draft
scuro::make is invoked by skeleton/Makefile. Its work is to point the
controlling package Makefile at the right files and directories, and to
ensure that rmarkdown/knitr spews its files in the right places.

skeleton.Rmd is required at the top level of the template directory by
rmarkdown. rmarkdown::draft gives a warning without it, and RStudio
won't list the template in the "New R Markdown" dialog. So I made a
README there. The actual samples are the {notes,script}-sample.Rmd
files.