lips: copy n paste
Chicago ellipses. Decided I wanted to avoid the dependency on my hacked
"lipskludge" package, but guess what? no other solution works. The
ellipsis package with option `chicago` does not perform well in
conjunction with pandoc's insistence on rendering ... as `\ldots{}` in
all cases. I thought I'd found a workaround for that, but it is defeated
by `amsmath` which changes the definition of `\dots` in a way that
ellipsis does not interact with well.

So instead: just pasting in the `lips` code, subtracting the stuff that
isn't needed. The only *change* to the code is to switch a `\newlet` to
`\let`. `\newlet` is a `moredefs` macro. moredefs has a bad interaction
with many packages. But we don't care either. It's a little obnoxious to
paste in a couple dozen lines of @@@@latex@@@ gunk but the TeX is only
an intermediate file and it saves having to carry around a .sty file.
Maybe I'll go back to a .sty eventually but for now this'll do.

We add a chicago-ellipses flag for this behavior.

unrelated: moved the biblatex name penalty stuff to only be executed if biblatex
loaded.