highlighting is tricky. chiaro_md -> chiaro_pdf
can't get syntax highlighting with a pandoc metadata key, so we have to
do some more work to read it from metadata and hand it to pandoc at the
make stage. `scuro::make` is renamed `render_pdf` and is meant to be
invoked one file at a time, really.

We factor out the knitr option fiddle into `scuro_knitr`.  This can now
be used in the article format `chiaro_pdf`. I learned that
`rmarkdown::render` is smart enough to use `latexmk`, so all the
footwork with intermediate markdown etc. can be skipped and we can just
make a PDF format.