citation_package; autodep improved
in order to get render to take biblatex into account, we have to tell it
a citation_package. So that's what we'll do. No more need for `biblatex:
true` in top level metadata, at least.

Then I missed the boat when autodep=TRUE no longer needed an explicit call to
dep_auto(). That was three years ago.

https://github.com/yihui/knitr-examples/commit/a2940f7bcc1b11a1e85cf30a299e654ac7110553

Anyway, remove dep_auto from the template Rmds. While I was at it,
folded in the chunk options to the format defaults rather than the
template Rmds.