further small fixes to render_pdf / scuro_md
We inject a metadata key `routput: true` to ensure that the LaTeX
template will have an `ROutput` definition even if there is no R code
(possible if we have chunks that are `echo=F` that give output).

N.B. `is.null(meta$routput)` can be false if the metadata contains

routput-options: ...

because `$` will partial-match it!

So we have to be more exacting and look at `[[...]]`.