a little less wacky on the title block
Arbitrarily switching on a font feature is really a pretty silly idea.
If someone is obsessive about that sort of thing they'll set the title
themselves anyway. This approach provides a plausible default, which can
be deactivated with manual-title: true.

For automatic titling, I go back to using the memoir titling
commands and maketitle. The key is that memoir actually does define
\maketitlehookb and so we have to make sure to redefine it or the
titling text gets overprinted on the author and date (seriously). The
date gets sent into the right header on the title page. The only thing
that can be configured here is the titlefont (and titlefontoptions), as demonstrated in paper.md.

Changed the metadata keys for adjusting headings after the title page.
These are now

[oe][lcr](head|foot)