"breaking" change: default slide_level is 1
scuro_md's `slide_level` parameter is used to set `NOTES_SLIDE_LEVEL`
when generating slides from notes (not from a script). The previous
level of 2 was not only counterintuitive but led to incorrect results
with the sample files from the template. 1 is now the default.

This change will however lead to incorrect results for any markdown written using the old default. The old behavior can be achieved with metadata `slide_level: 2` under `output: scuro::scuro_md:`.